@@ -1,6 +1,6 @@
 module github.com/yosebyte/miniclaw
 
-go 1.19
+go 1.25
 
 require (
 	github.com/go-telegram-bot-api/telegram-bot-api/v5 v5.5.1 // indirect
